requirements.txt
streamlit>=1.33.0,<2
pandas>=2.1.0,<3
matplotlib>=3.7.0,<4
google-generativeai>=0.7.0,<1
pillow>=10.0.0,<12
requests>=2.31.0,<3
python-dotenv>=1.0.0,<2
